Refactor builders to model.util package
