+++ src/test/java/seedu/address/logic/commands/AddStaffCommandTest.java
@@ -27,7 +27,7 @@ import seedu.address.model.drink.ReadOnlyDrinkCatalog;
 import seedu.address.model.person.Customer;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Staff;
-import seedu.address.testutil.StaffBuilder;
+import seedu.address.model.util.StaffBuilder;
 
 public class AddStaffCommandTest {
