+++ src/test/java/seedu/address/logic/commands/EditStaffCommandTest.java
@@ -26,8 +26,8 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Staff;
+import seedu.address.model.util.StaffBuilder;
 import seedu.address.testutil.EditStaffDescriptorBuilder;
-import seedu.address.testutil.StaffBuilder;
 
 /**
  * Contains integration tests (interaction with the Model) and unit tests for EditStaffCommand.
