+++ src/test/java/seedu/address/model/person/NameContainsKeywordsPredicateTest.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.testutil.PersonBuilder;
+import seedu.address.model.util.PersonBuilder;
 
 public class NameContainsKeywordsPredicateTest {
