+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -17,6 +17,7 @@ import java.util.List;
 
 import seedu.address.model.AddressBook;
 import seedu.address.model.person.Person;
+import seedu.address.model.util.PersonBuilder;
 
 /**
  * A utility class containing a list of {@code Person} objects to be used in tests.
